Version updates for the Android sample.
Update the Android sample to Google Play Services v11
and the latest support and compatibility library.
This change required the addition of an API key, which is
specified in the local.properties file and loaded through
the Gradle build configuration.
(These properties can be overriden with a local file in the home
directory.)

Change-Id: If3b1dbca0a168fa32b7a90e2b6abb7e809289880